+++ R/netest.R
@@ -29,8 +29,6 @@
 #'        details).
 #' @param set.control.ergm.ego Control arguments passed to \code{ergm.ego} (see
 #'        details).
-#' @param set.control.stergm Deprecated control argument of class
-#'        \code{control.stergm}; use \code{set.control.tergm} instead.
 #' @param set.control.tergm Control arguments passed to \code{tergm}
 #'        (see details).
 #' @param verbose If \code{TRUE}, print model fitting progress to console.
@@ -153,15 +151,9 @@
 netest <- function(nw, formation, target.stats, coef.diss, constraints,
                    coef.form = NULL, edapprox = TRUE,
                    set.control.ergm = control.ergm(),
-                   set.control.stergm = control.stergm(),
                    set.control.tergm = control.tergm(),
                    set.control.ergm.ego = control.ergm.ego(),
                    verbose = FALSE, nested.edapprox = TRUE, ...) {
-
-  if (!missing(set.control.stergm)) {
-    warning("set.control.stergm is deprecated and will be removed in a future
-             version; use set.control.tergm instead.")
-  }
 
   if (missing(constraints)) {
     constraints	<- trim_env(~.)
@@ -172,7 +164,7 @@ netest <- function(nw, formation, target.stats, coef.diss, constraints,
          call. = FALSE)
   }
   dissolution <- coef.diss$dissolution
-  if (coef.diss$coef.crude[1] == -Inf) {
+  if (coef.diss$duration[1] == 1) {
     is.tergm <- FALSE
   } else {
     is.tergm <- TRUE
@@ -184,30 +176,16 @@ netest <- function(nw, formation, target.stats, coef.diss, constraints,
 
   if (edapprox == FALSE) {
 
-    if (!missing(set.control.stergm)) {
-      fit <- stergm(nw,
-                    formation = formation,
-                    dissolution = dissolution,
-                    targets = "formation",
-                    target.stats = target.stats,
-                    offset.coef.form = coef.form,
-                    offset.coef.diss = coef.diss$coef.crude,
-                    constraints = constraints,
-                    estimate = "EGMME",
-                    eval.loglik = FALSE,
-                    control = set.control.stergm,
-                    verbose = verbose)
-    } else {
-      fit <- tergm(nw ~ Form(formation) + Persist(dissolution),
-                   targets = "formation",
-                   target.stats = target.stats,
-                   offset.coef = c(coef.form, coef.diss$coef.crude),
-                   constraints = constraints,
-                   estimate = "EGMME",
-                   eval.loglik = FALSE,
-                   control = set.control.tergm,
-                   verbose = verbose)
-    }
+    fit <- tergm(~Form(formation) + Persist(dissolution),
+                 basis = nw,
+                 targets = "formation",
+                 target.stats = target.stats,
+                 offset.coef = c(coef.form, coef.diss$coef.crude),
+                 constraints = constraints,
+                 estimate = "EGMME",
+                 eval.loglik = FALSE,
+                 control = set.control.tergm,
+                 verbose = verbose)
 
     coef.form <- fit # there is no longer a separate formation fit
     which_form <- which(grepl("^Form~", names(coef(fit))) |
@@ -228,14 +206,14 @@ netest <- function(nw, formation, target.stats, coef.diss, constraints,
     out$newnetwork <- as.network(fit$newnetwork)
     delete.network.attribute(out$newnetwork, "time")
     delete.network.attribute(out$newnetwork, "lasttoggle")
-    out$formula <- fit$formula
+    out$formula <- trim_env(fit$formula, keep = c("formation", "dissolution"))
 
   } else {
-    formation.nw <- nonsimp_update.formula(formation, nw ~ ., from.new = "nw")
 
     if (is(nw, "egor")) {
       # ergm.ego case
-      fit <- ergm.ego(formation.nw,
+      fit <- ergm.ego(formation,
+                      basis = nw,
                       popsize = 0,
                       constraints = constraints,
                       offset.coef = coef.form,
@@ -246,7 +224,8 @@ netest <- function(nw, formation, target.stats, coef.diss, constraints,
 
     } else {
       # ergm case
-      fit <- ergm(formation.nw,
+      fit <- ergm(formation,
+                  basis = nw,
                   target.stats = target.stats,
                   constraints = constraints,
                   offset.coef = coef.form,
@@ -265,12 +244,8 @@ netest <- function(nw, formation, target.stats, coef.diss, constraints,
         ## implement the edapprox by appending the dissolution model to the
         ## formation model and appending the relevant values to the vector of
         ## formation model coefficients
-        formula_addition <- append_rhs.formula(~., coef.diss$dissolution,
-                                               keep.onesided = TRUE,
-                                               env = environment(coef.diss$dissolution))
-
-        # the ... allows for copying via from.new
-        formation <- nonsimp_update.formula(formation, formula_addition, ...)
+        formation <- trim_env(~Passthrough(formation) + Passthrough(dissolution),
+                              keep = c("formation", "dissolution"))
         coef.form <- c(coef.form, -coef.diss$coef.form.corr)
       }
     }
@@ -286,7 +261,7 @@ netest <- function(nw, formation, target.stats, coef.diss, constraints,
     out$constraints <- constraints
     out$edapprox <- edapprox
     out$nested.edapprox <- nested.edapprox
-    out$newnetwork <- fit$newnetwork
+    out$newnetwork <- NVL(fit$newnetwork, fit$network)
     out$formula <- fit$formula
   }
 
@@ -346,14 +321,9 @@ diss_check <- function(formation, dissolution) {
     stop("Dissolution model is not a subset of formation model.", call. = FALSE)
   }
   if (!all(diss.terms[1, ] %in% c("edges", "nodemix",
-                                  "nodematch", "nodefactor"))) {
+                                  "nodematch"))) {
     stop("The only allowed dissolution terms are edges, nodemix,
-         nodematch and ", "nodefactor", call. = FALSE)
-  }
-  if (any(diss.terms[1, ] %in% c("nodefactor"))) {
-    warning("Support for dissolution models containing a nodefactor term is
-            deprecated, and will be removed in a future release.")
-    # TODO: remove functionality and deprecation message in future release
+         and nodematch", call. = FALSE)
   }
   if (any(matchpos != seq_len(ncol(diss.terms)))) {
     stop("Order of terms in the dissolution model does not correspond to the ",
@@ -380,7 +350,6 @@ diss_check <- function(formation, dissolution) {
 #' @param nested.edapprox Logical. If \code{edapprox = TRUE} the dissolution
 #'        model is an initial segment of the formation model (see details in
 #'        \code{\link{netest}}).
-#' @param ... Additional arguments passed to other functions.
 #'
 #' @details
 #' Fitting an ERGM is a computationally intensive process when the model
@@ -423,7 +392,7 @@ diss_check <- function(formation, dissolution) {
 #'}
 #'
 update_dissolution <- function(old.netest, new.coef.diss,
-                               nested.edapprox = TRUE, ...) {
+                               nested.edapprox = TRUE) {
 
   if (!inherits(old.netest, "netest")) {
     stop("old.netest must be an object of class netest", call. = FALSE)
@@ -449,17 +418,8 @@ update_dissolution <- function(old.netest, new.coef.diss,
     } else {
       ## remove the part of the formation model and coefficient vector
       ## corresponding to the old edapprox
-      old_diss_list <- list_rhs.formula(out$coef.diss$dissolution)
-
-      formation_list <- list_rhs.formula(out$formation)
-      formation_list <- formation_list[seq_len(length(formation_list) -
-                                                 length(old_diss_list))]
-
-      formation <- append_rhs.formula(~., formation_list,
-                                      env = environment(out$formation))
-      formation[[2]] <- NULL # remove the . on the LHS
-
-      out$formation <- formation
+      out$formation <- eval(quote(formation),
+                            envir = environment(out$formation))
       out$coef.form <-
         out$coef.form[seq_len(length(out$coef.form) -
                                 length(out$coef.diss$coef.form.corr))]
@@ -482,12 +442,12 @@ update_dissolution <- function(old.netest, new.coef.diss,
       ## implement new edapprox by appending the new dissolution model to the
       ## formation model and appending the relevant values to the vector of
       ## formation model coefficients
-      formula_addition <- append_rhs.formula(~., new.coef.diss$dissolution,
-                                             keep.onesided = TRUE,
-                                             env = environment(new.coef.diss$dissolution))
-      # the ... allows for copying via from.new
-      out$formation <- nonsimp_update.formula(out$formation,
-                                              formula_addition, ...)
+      # nolint start
+      formation <- out$formation
+      dissolution <- new.coef.diss$dissolution
+      # nolint end
+      out$formation <- trim_env(~Passthrough(formation) + Passthrough(dissolution),
+                                keep = c("formation", "dissolution"))
       out$coef.form <- c(out$coef.form, -new.coef.diss$coef.form.corr)
     }
   }