+++ R/netsim.R
@@ -4,9 +4,11 @@
 #' @description Simulates stochastic network epidemic models for infectious
 #'              disease.
 #'
-#' @param x Fitted network model object, as an object of class \code{netest}.
-#'        Alternatively, if restarting a previous simulation, may be an object
-#'        of class \code{netsim}.
+#' @param x If \code{control$start == 1}, either a fitted network model object
+#'        of class \code{netest} or a list of such objects. If
+#'        \code{control$start > 1}, an object of class \code{netsim}. When
+#'        multiple networks are used, the node sets (including network size
+#'        and nodal attributes) are assumed to be the same for all networks.
 #' @param param Model parameters, as an object of class \code{param.net}.
 #' @param init Initial conditions, as an object of class \code{init.net}.
 #' @param control Control settings, as an object of class
@@ -53,11 +55,12 @@
 #'        \code{\link{control.net}} and the
 #'        \href{http://www.epimodel.org/tut.html}{tutorials} for further
 #'        details.
-#'  \item \strong{network:} a list of \code{networkDynamic} objects,
-#'         one for each model simulation.
+#'  \item \strong{network:} a list of lists of \code{networkDynamic} or
+#'        \code{networkLite} objects, with one list of objects for each model
+#'        simulation.
 #' }
 #' If \code{control$raw.output == TRUE}: A list of the raw (pre-processed)
-#' \code{netsim} \code{dat} objects, for use in simulation continuation.
+#' \code{netsim_dat} objects, for use in simulation continuation.
 #'
 #' @references
 #' Jenness SM, Goodreau SM and Morris M. EpiModel: An R Package for Mathematical
@@ -171,7 +174,8 @@ netsim_validate_control <- function(control) {
       "verbose",
       ".checkpoint.keep",
       ".traceback.on.error",
-      ".dump.frame.on.error"
+      ".dump.frame.on.error",
+      "cumulative.edgelist"
     )
   )
 
@@ -179,6 +183,10 @@ netsim_validate_control <- function(control) {
     for (flag in control_default_bool[[val]])
       if (is.null(control[[flag]])) control[[flag]] <- as.logical(val)
   }
+
+  # truncate the cumulative edgelists to keep only active partnerships
+  if (is.null(control$truncate.el.cuml))
+    control$truncate.el.cuml <- 0
 
   if (is.null(control$start))
     control$start <- 1